+++ conf/uri.php
@@ -13,13 +13,6 @@ $conf = [
      */
     'routing' => [
         /*
-         * Default module for controllers.
-         *
-         * @var string
-         */
-        'module' => '',
-
-        /*
          * Default namespace for controllers.
          *
          * @var string